backend/requirements.txt
# Steplight Backend Requirements
flask>=3.0.0
flask-cors
google-genai
python-dotenv
Pillow
opencage
requests